// apps/exam-vault/src/generated/models/attempt-model.ts
import type { Exam } from './exam-model';

export const AttemptStatusKeyToLabel = {
  'StatusKey0': 'InProgress',
  'StatusKey1': 'Submitted',
  'StatusKey2': 'AutoCancelled',
  'StatusKey3': 'GradingPending',
  'StatusKey4': 'Completed'
} as const;
export type AttemptStatusKey = keyof typeof AttemptStatusKeyToLabel;

export interface Attempt {
  /**
   * @displayName Attempt ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Student Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  studentName: string;
  /**
   * @displayName Auto Score
   */
  autoScore?: number;
  /**
   * @displayName Exam
   * @validationRule Required for create/update operations
   */
  exam: Pick<Exam, 'id' | 'title'>;
  /**
   * @displayName Manual Score
   */
  manualScore?: number;
  /**
   * @displayName Started At
   * @validationRule Required for create/update operations
   */
  startedAt: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: AttemptStatusKey;
  /**
   * @displayName Student Email
   * @validationRule Required for create/update operations
   */
  studentEmail: string;
  /**
   * @displayName Student ID
   * @validationRule Required for create/update operations
   */
  studentID: string;
  /**
   * @displayName Submitted At
   */
  submittedAt?: string;
  /**
   * @displayName Total Possible
   * @validationRule Required for create/update operations
   */
  totalPossible: number;
  /**
   * @displayName Total Score
   */
  totalScore?: number;
  /**
   * @displayName Violation Count
   * @validationRule Required for create/update operations
   */
  violationCount: number;
}

export const _Attempt = 'Attempt' as const;