// apps/exam-vault/src/generated/models/exam-model.ts
export const ExamStatusKeyToLabel = {
  'StatusKey0': 'Draft',
  'StatusKey1': 'Published',
  'StatusKey2': 'Closed'
} as const;
export type ExamStatusKey = keyof typeof ExamStatusKeyToLabel;

export interface Exam {
  /**
   * @displayName Exam ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName Access Code
   * @validationRule Required for create/update operations
   */
  accessCode: string;
  /**
   * @displayName Allow Review
   * @validationRule Required for create/update operations
   */
  allowReview: boolean;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName Created By
   * @validationRule Required for create/update operations
   */
  createdBy: string;
  /**
   * @displayName Description
   */
  description?: string;
  /**
   * @displayName Duration (Minutes)
   * @validationRule Required for create/update operations
   */
  durationMinutes: number;
  /**
   * @displayName End Date Time
   */
  endDateTime?: string;
  /**
   * @displayName Passing Score (%)
   * @validationRule Required for create/update operations
   */
  passingScore: number;
  /**
   * @displayName Show Results Immediately
   * @validationRule Required for create/update operations
   */
  showResultsImmediately: boolean;
  /**
   * @displayName Start Date Time
   */
  startDateTime?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ExamStatusKey;
}

export const _Exam = 'Exam' as const;